Fixes #7
JSONP requests don't gracefully fail.  Since they're supposed to register a call back, if the request fails, it's unlikely that the script block that executes the callback wasn't created, and so it fails silently.

The easy option is to set a timeout, since in practice this request fails quickly, we can assume it has failed after abotu 2.5 seconds (probably less, but 2.5 seems safe).

There's a jquery-jsonp plugin that can be used to add this functionality.

 https://github.com/jaubourg/jquery-jsonp